pyproject.toml
[project]
name = "tg-finans"
version = "0.1.0"
description = "Telegram Personal Finance Tracker Bot"
requires-python = ">=3.9"
dependencies = [
    "python-telegram-bot>=21.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "openai>=1.50.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "aiosqlite>=0.20.0",
    "eval_type_backport>=0.2.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0",
    "ruff>=0.6.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py39"
line-length = 100
